Rename the driver module to dpctl_iface.
  - The name driver does not convey much, let us be explicit
    that the module is the interface to dpctl.
  - Move dpctl_functions.py into dpctl_iface.
