+++ numba_dppy/dpctl_iface/kernel_launch_ops.py
@@ -17,8 +17,8 @@ from numba.core.ir_utils import legalize_names
 
 from numba_dppy import numpy_usm_shared as nus
 from numba_dppy import utils
-from numba_dppy.driver import DpctlCAPIFnBuilder
-from numba_dppy.driver._helpers import numba_type_to_dpctl_typenum
+from numba_dppy.dpctl_iface import DpctlCAPIFnBuilder
+from numba_dppy.dpctl_iface._helpers import numba_type_to_dpctl_typenum
 
 
 class KernelLaunchOps:
